Split WorkflowActivate on multiple files
