+++ src/ggrc_workflows/assets/javascripts/index.js
@@ -17,7 +17,8 @@ import './controllers/approval_workflow_controller';
 import './components/workflow_clone';
 import './components/taskgroup_clone';
 import './components/end_cycle';
-import './components/workflow_active';
+import './components/workflow-activate';
+import './components/workflow-start-cycle';
 import './components/workflow_deactivate';
 import './components/repeat-on-button-wrapper';
 import './components/repeat-on-button';
